Split config classes into own file
